+++ example/firebase_cloud_messaging_interop_example.dart
@@ -1,20 +1,37 @@
 import 'package:firebase_cloud_messaging_interop/firebase_cloud_messaging_interop.dart';
 import 'dart:html';
 
-main() {
-  Messaging messaging = firebase.messaging();
 
-  messaging.usePublicVapidKey("YOUR_VAPID_KEY");
+class FCMService {
+  FCM fcm;
 
-  /// Ask for permission to send notification
-  Notification.requestPermission().then((permission) {
-    if (permission == 'granted') {
-      futureFromPromise(messaging.getToken().then((e) {
-        /// SEND TOKEN TO THE BACKEND SERVER
-      }));
-    }
-    else {
-      /// User doesn't want notification :(
-    }
-  });
+  String currentToken;
+
+  FCMService() {
+    fcm = FCM(publicVapidKey: "YourPublicVapidKey");
+
+    fcm.onMessage((e) {
+      /// You can access title, body and tag
+    });
+
+    fcm.onTokenRefresh(requestPermissionAndGetToken);
+  }
+
+  void requestPermissionAndGetToken() {
+
+    Notification.requestPermission().then((permission) {
+      if (permission == 'granted') {
+        fcm.getToken().then((e) {
+          currentToken = e;
+          /// SEND TOKEN TO THE BACKEND SERVER
+        });
+      }
+      else {
+        /// The user doesn't want notification :(
+      }
+    });
+  }
+
+  void deleteCurrentToken() => fcm.deleteToken(currentToken);
+
 }