+++ lib/firebase_cloud_messaging_interop.dart
@@ -1,3 +1,3 @@
 library firebase_cloud_messaging_interop;
 
-export 'src/firebase_cloud_messaging_interop_base.dart';
+export 'src/fcm_interface.dart';