+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Fixing multiple issues.
+- Adding an interface to make the use of the FCM interop easier.
+
 ## 1.0.0
 
 - Initial version